Reaf: 实现了创建用户的功能

+++ src/main/java/online/niuma/xiaocubao/pojo/BaseEntity.java
@@ -5,11 +5,9 @@ import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableField;
 import com.baomidou.mybatisplus.annotation.TableId;
 import io.swagger.annotations.ApiModelProperty;
-import io.swagger.models.auth.In;
 import lombok.Data;
 
 import java.time.LocalDateTime;
-import java.util.Date;
 
 /**
  * {@code @date} 2022/12/8
@@ -21,7 +19,7 @@ import java.util.Date;
 public abstract class BaseEntity {
 
     @TableId(value = "id", type = IdType.ASSIGN_ID)
-    private Integer id;
+    private Long id;
 
     @ApiModelProperty("创建时间")
     @TableField(fill = FieldFill.INSERT)
